refactor(maps): share the location-button and camera-seeding helpers
`handleLocationButtonClick`, `initialCameraPosition` and `toPosition` were
private to `SearchStopMap`, so a second full map surface would have to
reimplement the permission flow and the default camera - and would drift from it
the first time either changed.

Moves them to `MapLocationHelpers` unchanged. Behaviour on SearchStop is
identical; this only makes the behaviour reusable.
